refactor: 멤버 엔티티클래스를 별도의 패키지로 포함함
