+++ src/main/java/org/ktc2/cokaen/wouldyouin/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package org.ktc2.cokaen.wouldyouin.repository;
 
 import java.util.UUID;
-import org.ktc2.cokaen.wouldyouin.domain.Member;
+import org.ktc2.cokaen.wouldyouin.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
