+++ src/main/java/org/ktc2/cokaen/wouldyouin/service/MemberService.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.ktc2.cokaen.wouldyouin.controller.member.MemberCreateRequest;
 import org.ktc2.cokaen.wouldyouin.controller.member.MemberEditRequest;
 import org.ktc2.cokaen.wouldyouin.controller.member.MemberResponse;
-import org.ktc2.cokaen.wouldyouin.domain.Member;
+import org.ktc2.cokaen.wouldyouin.domain.member.Member;
 import org.ktc2.cokaen.wouldyouin.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
